#54: refactoring/reorganizing
