+++ src/routes/Routes.ts
@@ -7,6 +7,7 @@ import PrivacyPolicyPage from '../pages/PrivacyPolicyPage';
 import TermsOfServicePage from '../pages/TermsOfServicePage';
 import ManageExercises from '../pages/manage-exercises/ManageExercises';
 import EditWorkout from '../pages/edit-workout/EditWorkout';
+import ManageWorkouts from '../pages/manage-workouts/ManageWorkouts';
 
 const routes: IRoute[] = [
     {
@@ -42,6 +43,13 @@ const routes: IRoute[] = [
         exact: true,
         component: EditWorkout,
         name: 'Create Workout',
+        protected: true
+    },
+    {
+        path: '/workouts',
+        exact: true,
+        component: ManageWorkouts,
+        name: 'Manage Workouts',
         protected: true
     },
     {
